requirements.txt
pyyaml==5.4.1
requests==2.26.0
requests-cache==0.7.2

# for tests only:
xmldiff==2.4
